requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.7
cryptography==36.0.0
cycler==0.11.0
deprecation==2.1.0
distlib==0.3.1
distro==1.8.0
docopt==0.6.2
ecdsa==0.18.0
entrypoints==0.3
exceptiongroup==1.1.3
fastapi==0.104.0
filelock==3.0.12
fonttools==4.28.2
frozenlist==1.4.0
gotrue==2.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.4
importlib-resources==5.4.0
iniconfig==2.0.0
Jinja2==3.0.3
joblib==1.1.0
jose==1.0.0
jsonschema==4.2.1
kiwisolver==1.3.2
langchain==0.0.27
lightgbm==3.3.1
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.0
multidict==6.0.4
nba-api==1.1.11
numpy==1.21.4
openai==1.3.5
outcome==1.1.0
packaging==21.3
pandas==1.3.4
pbr==5.5.1
Pillow==8.4.0
pluggy==1.3.0
portpicker==1.5.0
postgrest==0.13.0
psutil==5.8.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2021.3
PyYAML==6.0.1
realtime==1.0.2
requests==2.26.0
rsa==4.9
scikit-learn==1.0.1
scipy==1.7.3
selenium==4.1.0
setuptools-scm==6.3.2
six==1.15.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==2.0.19
starlette==0.27.0
stevedore==3.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.1
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.8.0
urllib3==1.26.7
uvicorn==0.23.2
uvloop==0.19.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
watchfiles==0.21.0
websockets==11.0.3
wsproto==1.0.0
xgboost==1.5.1
yarl==1.9.2
zipp==3.6.0
